refactor: extract ServedVersionValidator to its own file
